+++ src/main/kotlin/de/moritzf/proxy/server/ProxyCallAttributes.kt
@@ -1,5 +1,6 @@
 package de.moritzf.proxy.server
 
+import de.moritzf.proxy.usage.UsageTracker
 import io.ktor.util.AttributeKey
 
 object ProxyCallAttributes {
@@ -7,4 +8,5 @@ object ProxyCallAttributes {
     val ADMIN_KEY_FINGERPRINT = AttributeKey<String>("adminKeyFingerprint")
     val KEY_NAME = AttributeKey<String>("keyName")
     val KEY_FINGERPRINT = AttributeKey<String>("keyFingerprint")
+    val USAGE_TRACKER = AttributeKey<UsageTracker>("usageTracker")
 }